+++ libs/gui/src/CDisplayWindowPlots.cpp
@@ -10,6 +10,7 @@
 #include "gui-precomp.h"  // Precompiled headers
 //
 #include <mrpt/config.h>
+#include <mrpt/core/lock_helper.h>
 #include <mrpt/gui/CDisplayWindowPlots.h>
 #include <mrpt/gui/WxSubsystem.h>
 #include <mrpt/gui/WxUtils.h>
@@ -177,6 +178,7 @@ void CWindowDialogPlots::OnChar(wxKeyEvent& event)
 		const int code = event.GetKeyCode();
 		const mrptKeyModifier mod = mrpt::gui::keyEventToMrptKeyModifier(event);
 
+		auto lck = mrpt::lockHelper(m_winPlots->m_mtx);
 		m_winPlots->m_keyPushedCode = code;
 		m_winPlots->m_keyPushedModifier = mod;
 		m_winPlots->m_keyPushed = true;