+++ libs/opengl/include/mrpt/opengl/CAssimpModel.h
@@ -72,7 +72,7 @@ class CAssimpModel : public CRenderizableShaderTriangles,
 	}
 	void enqueueForRenderRecursive(
 		const mrpt::opengl::TRenderMatrices& state, RenderQueue& rq,
-		bool wholeInView) const override;
+		bool wholeInView, bool is1stShadowMapPass) const override;
 	bool isCompositeObject() const override { return true; }
 	/** @} */