+++ libs/opengl/include/mrpt/opengl/CSetOfObjects.h
@@ -77,7 +77,7 @@ class CSetOfObjects : public CRenderizable
 	void renderUpdateBuffers() const override;
 	void enqueueForRenderRecursive(
 		const mrpt::opengl::TRenderMatrices& state, RenderQueue& rq,
-		bool wholeInView) const override;
+		bool wholeInView, bool is1stShadowMapPass) const override;
 	bool isCompositeObject() const override { return true; }
 	void freeOpenGLResources() override
 	{
@@ -131,9 +131,6 @@ class CSetOfObjects : public CRenderizable
 	bool traceRay(const mrpt::poses::CPose3D& o, double& dist) const override;
 
 	CRenderizable& setColor_u8(const mrpt::img::TColor& c) override;
-	CRenderizable& setColorR_u8(const uint8_t r) override;
-	CRenderizable& setColorG_u8(const uint8_t g) override;
-	CRenderizable& setColorB_u8(const uint8_t b) override;
 	CRenderizable& setColorA_u8(const uint8_t a) override;
 	bool contains(const CRenderizable::Ptr& obj) const;
 	mrpt::math::TBoundingBoxf internalBoundingBoxLocal() const override;