+++ libs/opengl/include/mrpt/opengl/CTextMessageCapable.h
@@ -8,11 +8,13 @@
    +------------------------------------------------------------------------+ */
 #pragma once
 
+#include <mrpt/containers/NonCopiableData.h>
 #include <mrpt/opengl/CRenderizable.h>
 #include <mrpt/opengl/CText.h>
 #include <mrpt/opengl/opengl_fonts.h>
 
 #include <map>
+#include <shared_mutex>
 
 namespace mrpt::opengl
 {
@@ -65,6 +67,8 @@ class CTextMessageCapable
 
 	struct TListTextMessages
 	{
+		mutable mrpt::containers::NonCopiableData<std::shared_mutex> mtx;
+
 		std::map<uint32_t, DataPerText> messages;
 
 		/** (re)generate all CText objects in the gl_text fields */