+++ libs/opengl/include/mrpt/opengl/DefaultShaders.h
@@ -22,6 +22,7 @@ namespace mrpt::opengl
 // inverse eye-distance for this reason.
 struct DefaultShaderID
 {
+	// Regular geometric elements:
 	static constexpr shader_id_t POINTS = 0;
 	static constexpr shader_id_t WIREFRAME = 1;
 	static constexpr shader_id_t TEXT = 2;
@@ -29,7 +30,17 @@ struct DefaultShaderID
 	static constexpr shader_id_t TEXTURED_TRIANGLES_LIGHT = 4;
 	static constexpr shader_id_t TRIANGLES_NO_LIGHT = 5;
 	static constexpr shader_id_t TEXTURED_TRIANGLES_NO_LIGHT = 6;
+
+	// Shadow generation 1st/2nd pass shaders:
+	static constexpr shader_id_t TRIANGLES_SHADOW_1ST = 10;
+	static constexpr shader_id_t TRIANGLES_SHADOW_2ND = 11;
+	static constexpr shader_id_t TEXTURED_TRIANGLES_SHADOW_1ST = 12;
+	static constexpr shader_id_t TEXTURED_TRIANGLES_SHADOW_2ND = 13;
+
+	// Special effects:
 	static constexpr shader_id_t SKYBOX = 20;
+	static constexpr shader_id_t DEBUG_TEXTURE_TO_SCREEN = 30;
+	static constexpr shader_id_t NONE = 31;	 //!< Skip rendering
 };
 
 /** Loads a set of OpenGL Vertex+Fragment shaders from the default library
@@ -39,5 +50,11 @@ struct DefaultShaderID
  * \ingroup mrpt_opengl_grp
  */
 Program::Ptr LoadDefaultShader(const shader_id_t id);
+
+// Use GL_TEXTURE0 for diffuse map material textures:
+static constexpr int MATERIAL_DIFFUSE_TEXTURE_UNIT = 0;
+
+// Use GL_TEXTURE1 for shadow map:
+static constexpr int SHADOW_MAP_TEXTURE_UNIT = 1;
 
 }  // namespace mrpt::opengl