+++ libs/opengl/src/DefaultShaders.cpp
@@ -23,10 +23,14 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 	// Vertex shader:
 	const char* vertex_shader = nullptr;
 	const char* fragment_shader = nullptr;
+	const char* fragShaderIncludes = nullptr;
 	std::vector<std::string> attribs, uniforms;
 
 	switch (id)
 	{
+		// ==============================
+		// Regular geometric elements
+		// ==============================
 		case DefaultShaderID::POINTS:
 			vertex_shader =
 #include "../shaders/points.v.glsl"
@@ -44,7 +48,6 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 			attribs = {"position", "vertexColor"};
 			break;
 
-			// ==============================
 		case DefaultShaderID::WIREFRAME:
 			vertex_shader =
 #include "../shaders/wireframe.v.glsl"
@@ -56,7 +59,6 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 			attribs = {"position", "vertexColor"};
 			break;
 
-			// ==============================
 		case DefaultShaderID::TRIANGLES_LIGHT:
 			vertex_shader =
 #include "../shaders/triangles-light.v.glsl"
@@ -64,12 +66,13 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 			fragment_shader =
 #include "../shaders/triangles-light.f.glsl"
 				;
-			uniforms = {"p_matrix",		   "v_matrix",		"m_matrix",
-						"light_diffuse",   "light_ambient", "light_specular",
-						"light_direction", "cam_position",	"materialSpecular"};
+			uniforms = {"p_matrix",		   "v_matrix",		  "m_matrix",
+						"light_diffuse",   "light_ambient",	  "light_specular",
+						"light_color",	   "light_direction", "cam_position",
+						"materialSpecular"};
 			attribs = {"position", "vertexColor", "vertexNormal"};
 			break;
-			// ==============================
+
 		case DefaultShaderID::TRIANGLES_NO_LIGHT:
 			vertex_shader =
 #include "../shaders/triangles-no-light.v.glsl"
@@ -80,7 +83,7 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 			uniforms = {"pmv_matrix"};
 			attribs = {"position", "vertexColor"};
 			break;
-			// ==============================
+
 		case DefaultShaderID::TEXTURED_TRIANGLES_LIGHT:
 			vertex_shader =
 #include "../shaders/textured-triangles-light.v.glsl"
@@ -88,13 +91,13 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 			fragment_shader =
 #include "../shaders/textured-triangles-light.f.glsl"
 				;
-			uniforms = {"p_matrix",		   "v_matrix",		  "m_matrix",
-						"light_diffuse",   "light_ambient",	  "light_specular",
-						"cam_position",	   "light_direction", "textureSampler",
-						"materialSpecular"};
+			uniforms = {"p_matrix",		  "v_matrix",		 "m_matrix",
+						"light_diffuse",  "light_ambient",	 "light_specular",
+						"light_color",	  "cam_position",	 "light_direction",
+						"textureSampler", "materialSpecular"};
 			attribs = {"position", "vertexUV", "vertexNormal"};
 			break;
-			// ==============================
+
 		case DefaultShaderID::TEXTURED_TRIANGLES_NO_LIGHT:
 			vertex_shader =
 #include "../shaders/textured-triangles-no-light.v.glsl"
@@ -105,7 +108,7 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 			uniforms = {"pmv_matrix", "textureSampler"};
 			attribs = {"position", "vertexUV"};
 			break;
-			// ==============================
+
 		case DefaultShaderID::TEXT:
 			vertex_shader =
 #include "../shaders/text.v.glsl"
@@ -116,7 +119,72 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 			uniforms = {"p_matrix", "mv_matrix"};
 			attribs = {"position", "vertexColor"};
 			break;
-			// ==============================
+
+		// ===========================================
+		// Shadow generation 1st/2nd pass shaders
+		// ===========================================
+		case DefaultShaderID::TRIANGLES_SHADOW_1ST:
+			vertex_shader =
+#include "../shaders/triangles-shadow-1st.v.glsl"
+				;
+			fragment_shader =
+#include "../shaders/triangles-shadow-1st.f.glsl"
+				;
+			uniforms = {"m_matrix", "light_pv_matrix"};
+			attribs = {"position"};
+			break;
+
+		case DefaultShaderID::TRIANGLES_SHADOW_2ND:
+			vertex_shader =
+#include "../shaders/triangles-shadow-2nd.v.glsl"
+				;
+			fragment_shader =
+#include "../shaders/triangles-shadow-2nd.f.glsl"
+				;
+			fragShaderIncludes =
+#include "../shaders/shadow-calculation.f.glsl"
+				;
+			uniforms = {
+				"p_matrix",		   "v_matrix",		 "m_matrix",
+				"light_pv_matrix", "shadowMap",		 "light_diffuse",
+				"light_ambient",   "light_specular", "light_direction",
+				"light_color",	   "cam_position",	 "materialSpecular"};
+			attribs = {"position", "vertexColor", "vertexNormal"};
+			break;
+
+			// 1st stage is the same for textured and non-textured triangles:
+		case DefaultShaderID::TEXTURED_TRIANGLES_SHADOW_1ST:
+			vertex_shader =
+#include "../shaders/triangles-shadow-1st.v.glsl"
+				;
+			fragment_shader =
+#include "../shaders/triangles-shadow-1st.f.glsl"
+				;
+			uniforms = {"m_matrix", "light_pv_matrix"};
+			attribs = {"position"};
+			break;
+
+		case DefaultShaderID::TEXTURED_TRIANGLES_SHADOW_2ND:
+			vertex_shader =
+#include "../shaders/textured-triangles-shadow-2nd.v.glsl"
+				;
+			fragment_shader =
+#include "../shaders/textured-triangles-shadow-2nd.f.glsl"
+				;
+			fragShaderIncludes =
+#include "../shaders/shadow-calculation.f.glsl"
+				;
+			uniforms = {"p_matrix",		   "v_matrix",		 "m_matrix",
+						"light_pv_matrix", "shadowMap",		 "light_diffuse",
+						"light_ambient",   "light_specular", "light_direction",
+						"light_color",	   "cam_position",	 "materialSpecular",
+						"textureSampler"};
+			attribs = {"position", "vertexNormal", "vertexUV"};
+			break;
+
+		// ===========================================
+		// Special effects
+		// ===========================================
 		case DefaultShaderID::SKYBOX:
 			vertex_shader =
 #include "../shaders/skybox.v.glsl"
@@ -126,6 +194,17 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 				;
 			uniforms = {"p_matrix", "v_matrix_no_translation", "skybox"};
 			attribs = {"position"};
+			break;
+
+		case DefaultShaderID::DEBUG_TEXTURE_TO_SCREEN:
+			vertex_shader =
+#include "../shaders/debug_show_texture.v.glsl"
+				;
+			fragment_shader =
+#include "../shaders/debug_show_texture.f.glsl"
+				;
+			uniforms = {"textureId"};
+			attribs = {"position", "vertexUV"};
 			break;
 
 		default:
@@ -161,20 +240,26 @@ Program::Ptr mrpt::opengl::LoadDefaultShader(const shader_id_t id)
 	std::string errMsgs;
 	std::vector<Shader> lstShaders;
 	lstShaders.resize(2);
-	if (!lstShaders[0].compile(GL_VERTEX_SHADER, vertex_shader, errMsgs))
+	if (!lstShaders[0].compile(GL_VERTEX_SHADER, {vertex_shader}, errMsgs))
 	{
 		THROW_EXCEPTION_FMT(
-			"Error compiling GL_VERTEX_SHADER:\n%s", errMsgs.c_str());
+			"Error compiling GL_VERTEX_SHADER (%s):\n%s", vertex_shader,
+			errMsgs.c_str());
 	}
-	if (!lstShaders[1].compile(GL_FRAGMENT_SHADER, fragment_shader, errMsgs))
+	if (!lstShaders[1].compile(
+			GL_FRAGMENT_SHADER,
+			{fragShaderIncludes ? fragShaderIncludes : "", fragment_shader},
+			errMsgs))
 	{
 		THROW_EXCEPTION_FMT(
-			"Error compiling GL_FRAGMENT_SHADER:\n%s", errMsgs.c_str());
+			"Error compiling GL_FRAGMENT_SHADER (%s):\n%s", fragment_shader,
+			errMsgs.c_str());
 	}
 	if (!shader->linkProgram(lstShaders, errMsgs))
 	{
 		THROW_EXCEPTION_FMT(
-			"Error linking Opengl Shader programs:\n%s", errMsgs.c_str());
+			"Error linking Opengl Shader programs (%s, %s):\n%s", vertex_shader,
+			fragment_shader, errMsgs.c_str());
 	}
 
 #if 0